small refactor
- rename `PredicateHasLabel` to `PredicateHasLabelWithValue`
- move `listOrganizationsAsReconcileRequests` to `util.go`

On-behalf-of: @SAP mikolaj.kucinski@sap.com
